refactor(client): relocate Pill primitive from radar to ui
Pill is a generic span primitive, not radar-specific. Move it to its
cross-cutting home so other features can reuse it without a radar dependency,
and update the three radar importers.
